Add logic for requests with client key on the header

+++ Sources/App/Controllers/GenericBaseCRUDController.swift
@@ -1,7 +1,7 @@
-import Fluent
 import Vapor
+import Fluent
 
-enum BaseCRUDOperations: Int {
+enum BaseCRUDControllerOperations: Int {
     case all
     case get
     case add
@@ -12,7 +12,7 @@ enum BaseCRUDOperations: Int {
 
 class BaseCRUDController {
         
-    var validOperations: [BaseCRUDOperations] = [.get]
+    var validOperations: [BaseCRUDControllerOperations] = [.get]
     
     static func allRecords<T: Model>(from:T.Type, req: Request) throws -> EventLoopFuture<[T]> {
         return try DatabaseUtils.CRUD.all(from: from.self, using: req.db)
@@ -24,40 +24,45 @@ class BaseCRUDController {
     }
 
     static func getRecord<T: Model>(from:T.Type, req: Request) throws -> EventLoopFuture<T> {
-        guard let requesParamUUID = req.requesParamUUID, let id = requesParamUUID as? T.IDValue  else { throw Abort(.badRequest) }
+        guard let requesParamUUID = req.paramUUID, let id = requesParamUUID as? T.IDValue  else { throw Abort(.badRequest) }
         return try DatabaseUtils.CRUD.get(from, id: id, using: req.db)
     }
     
     static func deleteRecord<T: Model>(from:T.Type, req: Request) throws -> EventLoopFuture<HTTPStatus> {
-        guard let requesParamUUID = req.requesParamUUID, let id = requesParamUUID as? T.IDValue  else { throw Abort(.badRequest) }
+        guard let requesParamUUID = req.paramUUID, let id = requesParamUUID as? T.IDValue  else { throw Abort(.badRequest) }
         return try DatabaseUtils.CRUD.delete(from, id: id, using: req.db).transform(to: .ok)
     }
 }
 
 
-class GenericBaseCRUDController<T: Model & RoutablePathProtocol & Content>: BaseCRUDController, RouteCollection {
+class GenericBaseCRUDController<T: Model & RoutableCRUDPath & Content>: BaseCRUDController, RouteCollection {
     func boot(routes: RoutesBuilder) throws {
-        let routeBasePath = T.path
-        let mainGroup = routes.grouped(routeBasePath)
+        let pathComponent = T.pathComponent
+        let mainGroup = routes.grouped(pathComponent)
 
         guard validOperations.count > 0 else {
-            DevTools.Logs.log(error: "No valid operations [\(BaseCRUDOperations.self)] where given. Ignored", app: nil)
+            DevTools.Logs.log(error: "No valid operations [\(BaseCRUDControllerOperations.self)] where given. Ignored", app: nil)
             return
         }
         
-        DevTools.Logs.log(message: "Will create routes at \\\(routeBasePath)", app: nil)
+        DevTools.Logs.log(message: "Will create routes at \\\(pathComponent)", app: nil)
 
-        if validOperations.contains(.get) || validOperations.contains(.all) {
+        let hasGetPermission = validOperations.contains(.get) || validOperations.contains(.all)
+        let hasSavePermission = validOperations.contains(.add) || validOperations.contains(.all)
+
+        if hasGetPermission {
             // HTTP GET : SERVER_PATH:PORT/INITIAL_ROUTE
             mainGroup.get { (req) -> EventLoopFuture<[T]> in
-                try Self.allRecords(from: T.self, req: req)
+                guard req.isValid else { throw Abort(.unauthorized) }
+                return try Self.allRecords(from: T.self, req: req)
             }
         }
 
-        if validOperations.contains(.add) || validOperations.contains(.all) {
+        if hasSavePermission {
             // HTTP POST : SERVER_PATH:PORT/INITIAL_ROUTE
             mainGroup.post { (req) -> EventLoopFuture<T> in
-                try Self.saveRecord(from: T.self, req: req)
+                guard req.isValid else { throw Abort(.unauthorized) }
+                return try Self.saveRecord(from: T.self, req: req)
             }
         }
 
@@ -65,7 +70,8 @@ class GenericBaseCRUDController<T: Model & RoutablePathProtocol & Content>: Base
             mainGroup.group("operations", ":\(DevTools.Strings.pathConventionId)") {
                 // HTTP GET : SERVER_PATH:PORT/INITIAL_ROUTE/operations/someId
                 some in some.get{ (req) -> EventLoopFuture<T> in
-                    try Self.getRecord(from: T.self, req: req)
+                    guard req.isValid else { throw Abort(.unauthorized) }
+                    return try Self.getRecord(from: T.self, req: req)
                 }
             }
         }
@@ -74,7 +80,8 @@ class GenericBaseCRUDController<T: Model & RoutablePathProtocol & Content>: Base
             mainGroup.group("operations", ":\(DevTools.Strings.pathConventionId)") {
                 // HTTP DELETE : SERVER_PATH:PORT/INITIAL_ROUTE/operations/someId
                 some in some.delete{ (req) -> EventLoopFuture<HTTPStatus> in
-                    try Self.deleteRecord(from: T.self, req: req)
+                    guard req.isValid else { throw Abort(.unauthorized) }
+                    return try Self.deleteRecord(from: T.self, req: req)
                 }
             }
         }
